add: mensagens de log e dtos

+++ 03-spring-web/PessoaAPI-Aula/src/main/java/br/com/dbc/vemser/pessoaapi/service/PessoaService.java
@@ -1,52 +1,65 @@
 package br.com.dbc.vemser.pessoaapi.service;
 
+import br.com.dbc.vemser.pessoaapi.dto.CreatePessoaDto;
+import br.com.dbc.vemser.pessoaapi.dto.ResponsePessoaDto;
 import br.com.dbc.vemser.pessoaapi.entity.Pessoa;
 import br.com.dbc.vemser.pessoaapi.exception.RegraDeNegocioException;
+import br.com.dbc.vemser.pessoaapi.mapper.PessoaMapper;
 import br.com.dbc.vemser.pessoaapi.repository.PessoaRepository;
 
+import br.com.dbc.vemser.pessoaapi.utils.PropertieReader;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@Slf4j
 public class PessoaService {
 
     private final PessoaRepository pessoaRepository;
+    private final PropertieReader propertieReader;
 
-    public PessoaService(PessoaRepository pessoaRepository){
+    public PessoaService(PessoaRepository pessoaRepository, PropertieReader propertieReader){
         this.pessoaRepository = pessoaRepository;
+        this.propertieReader = propertieReader;
     }
 
-    public Pessoa create(Pessoa pessoa) throws Exception {
-        if (StringUtils.isBlank(pessoa.getNome())) {
-            throw new RegraDeNegocioException("Nome inválido!");
-        }
-        return pessoaRepository.create(pessoa);
+    public ResponsePessoaDto create(CreatePessoaDto pessoa) throws Exception {
+        return PessoaMapper.pessoaToPessoaResponseDto(
+                pessoaRepository.create(
+                        PessoaMapper.createPessoaDtoToPessoa(pessoa)));
     }
 
-    public List<Pessoa> list() throws RegraDeNegocioException {
-        return pessoaRepository.list();
+    public List<ResponsePessoaDto> list() throws RegraDeNegocioException {
+        return pessoaRepository.list().stream()
+                .map(PessoaMapper::pessoaToPessoaResponseDto)
+                .toList();
     }
 
-    public Pessoa update(Integer id,
-                         Pessoa pessoaAtualizar) throws Exception {
+    public ResponsePessoaDto update(Integer id,
+                         CreatePessoaDto pessoaAtualizar) throws Exception {
+
         Pessoa pessoaRecuperada = getPessoa(id);
 
         pessoaRecuperada.setCpf(pessoaAtualizar.getCpf());
         pessoaRecuperada.setNome(pessoaAtualizar.getNome());
         pessoaRecuperada.setDataNascimento(pessoaAtualizar.getDataNascimento());
 
-        return pessoaRecuperada;
+
+        return PessoaMapper.pessoaToPessoaResponseDto(pessoaRecuperada);
     }
 
     public void delete(Integer id) throws Exception {
+        if (!propertieReader.getAdmin()) throw new RegraDeNegocioException("Não é possível deletar pessoas sem ser o administrador");
         Pessoa pessoaRecuperada = getPessoa(id);
         pessoaRepository.delete(pessoaRecuperada);
     }
 
-    public List<Pessoa> listByName(String nome) throws Exception{
-        return pessoaRepository.listByName(nome);
+    public List<ResponsePessoaDto> listByName(String nome) throws Exception{
+        return pessoaRepository.listByName(nome).stream().map(PessoaMapper::pessoaToPessoaResponseDto)
+                .toList();
     }
 
     public Pessoa getPessoa(Integer id) throws Exception {
